hashMap for collisionDetection and outOfWindow detection
	new file:   TowerDefenceANTSPrototype2/src/basics/ANTSCollisionDetection.java
	modified:   TowerDefenceANTSPrototype2/src/basics/ANTSDriver.java
	modified:   TowerDefenceANTSPrototype2/src/basics/ANTSFactory.java
	modified:   TowerDefenceANTSPrototype2/src/basics/ANTSWindow.java
	modified:   TowerDefenceANTSPrototype2/src/controllers/ANTSAbstractController.java
	modified:   TowerDefenceANTSPrototype2/src/controllers/ANTSGameController.java
	modified:   TowerDefenceANTSPrototype2/src/interfaces/ANTSIController.java
	modified:   TowerDefenceANTSPrototype2/src/interfaces/ANTSIModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSAbstractModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSGridModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSSimpleMediumModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSSimpleRayLightModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSSimpleSourceLightModel.java

+++ TowerDefenceANTSPrototype2/src/models/ANTSGridModel.java
@@ -84,14 +84,14 @@ public class ANTSGridModel extends ANTSAbstractModel implements ANTSIModel
 		return this.yOffset;
 	}
 	
-	public double getPosX()
+	public int getPosX()
 	{
 		return this.xOffset;
 	}
 	
-	public double getPosY()
+	public int getPosY()
 	{
-		return this.yOffset-this.cellHeight/2;
+		return (int) (this.yOffset-this.cellHeight/2);
 	}
 	
 	public double getHeight()
